run-playbook added checks

+++ run-playbook.sh
@@ -1,9 +1,24 @@
-#!/bin/sh
+#!/bin/bash
 
-unset AHOSTNAME
-read -p "Enter hostname: " AHOSTNAME
+if [ -n "$1" ]; then
+  AHOSTNAME=$1
+else
+  unset AHOSTNAME
+  read -p "Enter hostname: " AHOSTNAME
+fi
 echo "$AHOSTNAME" >inventory.tmp
 
-ansible-playbook -i inventory.tmp playbook.yaml
+ping $AHOSTNAME -c 1
+if [ $? != 0 ]; then
+  echo "Cannot reach $AHOSTNAME"
+  exit 1
+fi
+
+ansible -i inventory.tmp -m ping -u pi -b all
+if [ $? != 0 ]; then
+  echo "Cannot run ansible on host $AHOSTNAME"
+  exit 1
+fi
+ansible-playbook -i inventory.tmp -u pi playbook.yaml
 
 rm inventory.tmp
